+++ packages/nyris-webapp/src/components/pre-filter/index.tsx
@@ -132,7 +132,6 @@ function PreFilterComponent(props: Props) {
         },
       ];
       dispatch(updateStatusLoading(true));
-
       find({
         image: requestImage?.canvas as HTMLCanvasElement,
         settings,