+++ packages/nyris-webapp/src/components/results/ItemResult.tsx
@@ -1,6 +1,6 @@
 import { Button, Grid, Tooltip, Typography } from '@material-ui/core';
 import ChevronRightOutlinedIcon from '@material-ui/icons/ChevronRightOutlined';
-import IconOpenLink from 'common/assets/icons/Union.svg';
+import { ReactComponent as IconOpenLink } from 'common/assets/icons/Union.svg';
 import { ReactComponent as IconShare } from 'common/assets/icons/Fill.svg';
 import { ReactComponent as IconDisLike } from 'common/assets/icons/icon_dislike.svg';
 import { ReactComponent as IconLike } from 'common/assets/icons/icon_like.svg';
@@ -244,14 +244,14 @@ function ItemResult(props: Props) {
               justifyContent: 'space-between',
               flexDirection: 'column',
               gridGap: 8,
-              color: '#2B2C46',
+              color: settings.theme.mainTextColor || '#2B2C46',
             }}
           >
             {settings.CTAButtonText && (
               <Typography
                 className="text-f12 max-line-1 fw-700"
                 style={{
-                  color: '#2B2C46',
+                  color: settings.theme.mainTextColor || '#2B2C46',
                   marginTop: 8,
                 }}
               >
@@ -264,8 +264,8 @@ function ItemResult(props: Props) {
                 justifyContent: 'space-between',
                 flexDirection: 'row',
                 gridGap: 8,
-                color: '#2B2C46',
-                marginTop: 8,
+                color: settings.theme.mainTextColor || '#2B2C46',
+                marginTop: settings.CTAButtonText ? 0 : 8,
               }}
             >
               <Tooltip
@@ -277,7 +277,7 @@ function ItemResult(props: Props) {
                 <Typography
                   className="text-f12 max-line-1 fw-400"
                   style={{
-                    color: '#2B2C46',
+                    color: settings.theme.mainTextColor || '#2B2C46',
                   }}
                 >
                   {truncateString(
@@ -294,7 +294,7 @@ function ItemResult(props: Props) {
                   <Typography
                     className="text-f12 max-line-1 fw-400"
                     style={{
-                      color: '#2B2C46',
+                      color: settings.theme?.mainTextColor || '#2B2C46',
                     }}
                   >
                     <span
@@ -316,15 +316,17 @@ function ItemResult(props: Props) {
                 justifyContent: 'space-between',
                 flexDirection: 'row',
                 gridGap: 8,
-                color: '#2B2C46',
+                color: settings.theme.mainTextColor || '#2B2C46',
               }}
             >
               {(brand || settings.brandName) && (
                 <ProductAttribute
                   title={t('Brand')}
                   value={brand || settings.brandName}
-                  padding="4px 8px"
+                  padding={settings.theme.brandFieldPadding || '4px 8px'}
                   width={{ xs: '49%' }}
+                  backgroundColor={settings.theme.brandFieldBackground}
+                  isTitleVisible={settings.isBrandNameTitleVisible}
                 />
               )}
 
@@ -345,7 +347,7 @@ function ItemResult(props: Props) {
               display: 'flex',
               justifyContent: 'space-between',
               gridGap: 10,
-              color: '#2B2C46',
+              color: settings.theme.mainTextColor || '#2B2C46',
               marginTop: 8,
             }}
           >
@@ -386,7 +388,7 @@ function ItemResult(props: Props) {
                 boxShadow: '-2px 2px 4px rgba(170, 171, 181, 0.5)',
                 // marginBottom: 22,
                 height: 40,
-                background: '#2B2C46',
+                background: settings.theme.secondaryCTAButtonColor || '#2B2C46',
                 borderRadius: 4,
                 padding: '0px 8px',
                 marginTop: '8px',
@@ -447,7 +449,7 @@ function ItemResult(props: Props) {
                   boxShadow: '-2px 2px 4px rgba(170, 171, 181, 0.5)',
                   // marginBottom: 22,
                   height: 40,
-                  background: settings.theme?.primaryColor,
+                  background: settings.theme?.CTAButtonColor || settings.theme?.primaryColor,
                   borderRadius: 4,
                   padding: '0px 8px',
                   marginTop: '8px',
@@ -473,11 +475,12 @@ function ItemResult(props: Props) {
                   }}
                 >
                   <Typography
-                    className="text-white max-line-2"
+                    className="max-line-2"
                     style={{
                       overflow: 'hidden',
                       textOverflow: 'ellipsis',
                       fontWeight: 500,
+                      color: settings.theme?.CTAButtonTextColor || '#FFFFFF',
                       fontSize: '12px',
                       letterSpacing: '0.27px',
                       wordBreak: 'break-all',
@@ -489,7 +492,7 @@ function ItemResult(props: Props) {
                     {truncateString(dataItem[settings.field.productName], 45)}
                   </Typography>
                   {!isMobile && ctaLink && (
-                    <img src={IconOpenLink} alt="more-info" width={16} />
+                    <IconOpenLink fill={settings.theme?.CTAButtonTextColor || '#FFFFFF'} width={16} />
                   )}
                 </div>
               </div>
@@ -500,7 +503,7 @@ function ItemResult(props: Props) {
                 boxShadow: '-2px 2px 4px rgba(170, 171, 181, 0.5)',
                 // marginBottom: 22,
                 height: 40,
-                background: settings.theme?.primaryColor,
+                background: settings.theme?.CTAButtonColor || settings.theme?.primaryColor,
                 borderRadius: 4,
                 padding: '0px 8px',
                 marginTop: '8px',
@@ -526,11 +529,12 @@ function ItemResult(props: Props) {
                 }}
               >
                 <Typography
-                  className="text-white max-line-2"
+                  className="max-line-2"
                   style={{
                     overflow: 'hidden',
                     textOverflow: 'ellipsis',
                     fontWeight: 500,
+                    color: settings.theme?.CTAButtonTextColor || '#FFFFFF',
                     fontSize: '12px',
                     letterSpacing: '0.27px',
                     wordBreak: 'break-all',
@@ -542,7 +546,7 @@ function ItemResult(props: Props) {
                   {settings.CTAButtonText}
                 </Typography>
                 {!isMobile && ctaLink && (
-                  <img src={IconOpenLink} alt="more-info" width={16} />
+                  <IconOpenLink fill={settings.theme?.CTAButtonTextColor || '#FFFFFF'} width={16} />
                 )}
               </div>
             </div>