+++ packages/nyris-webapp/src/types.ts
@@ -53,6 +53,7 @@ export interface AppSettings extends NyrisAPISettings {
   appTitle?: string;
   auth0: Auth0Settings;
   brandName?: string;
+  isBrandNameTitleVisible?: boolean;
   cadenas?: Cadenas;
   CTAButtonText?: string;
   secondaryCTAButtonText?: string;
@@ -78,6 +79,7 @@ export interface AppSettings extends NyrisAPISettings {
   theme: SearchSuiteSettings;
   visualSearchFilterKey?: string;
   warehouseVariant?: boolean;
+  shouldUseUserMetadata?: boolean;
 }
 
 export interface SearchSuiteSettings {
@@ -88,6 +90,12 @@ export interface SearchSuiteSettings {
   mobileFooterImageColor?: string;
   primaryColor?: string;
   secondaryColor?: string;
+  secondaryCTAButtonColor?: string;
+  CTAButtonColor?: string;
+  CTAButtonTextColor?: string;
+  mainTextColor?: string;
+  brandFieldBackground?: string;
+  brandFieldPadding?: string;
 }
 
 export type AppState = {