+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.52 - 2024-04-18
+
+### Fix
+- react webapp not publish build
+
+### Added
+- use user metadata from auth0
+- changing secondary CTA button style
+
 ## 0.3.51 - 2024-04-17
 
 ### Added